added browser()
@mmmmtoasty19 , yes, that's the basic idea. I haven't used `enquo` with `!!` before, so I am not sure that's better that aes_string, but it seems it might be (it's more moderns, that's for sure).

Your aes(group = ) need to be mapped onto the crossproduct of race_ethnicity and sex to display properly.

Also, in case you don't know about it, use `browser()` fundtion to debug and study the behavior of elemnts within the function. If you define a function with a `browser()` call inside, iit will halt when function is executed and take you to the environment in that point of the function.